% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-data.R
\docType{data}
\name{distance}
\alias{distance}
\title{Distances in Sheffield}
\format{A data frame with twenty observations on two variables}
\source{
Gilchrist, W. (1984). \emph{Statistical Modelling}. Chichester: John Wiley \& Sons, 5.
}
\usage{
distance
}
\description{
Data of the distance by road and the straight line (Euclidian) distance between twenty different
pairs of points in Sheffield
}
\details{
Dataset number 115, page 87. Dataset last updated on CRC-Press on 30 January 2008.
}
\references{
Hand, D., Daly, F., McConway, K., Lunn, D., Ostrowski, E. (1993). A Handbook of Small Data Sets. New York: Chapman and Hall/CRC.
}
\keyword{datasets}
