% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KRS-data.R
\docType{data}
\name{KRS}
\alias{KRS}
\title{Khayelitsha Rodent Study 2017-2018}
\format{A tibble with 222 observations over 266 variables}
\source{
DataFirst, University of Cape Town
}
\usage{
KRS
}
\description{
The main objective of the Khayelitsha Rodent Study (KRS) was to explore the nature of rodent infestation and attitudes to rodent control and poison use in Site C Khayelitisha, Cape Town, South Africa. The KRS originated in response to the controversy over a Public Works Programme (PWP) run by the Khayeltisha Environmental Health Unit in which previously unemployed people were hired to set cage traps in people’s homes and then drown the captured rats. The Society for the Prevention of Cruelty to Animals (SPCA) threatened legal action (because drowning animals is illegal) and the PWP resorted to using poison to destroy the rats.
}
\details{
For more detials see the DataFirst page https://www.datafirst.uct.ac.za/dataportal/index.php/catalog/725/
}
\references{
Centre for Social Science Research and the Institute for Communities and Wildlife, University of Cape Town. 2018. Khayelitsha Rodent Study 2017-2018 [Dataset]. Version 1. Cape Town: CSSR and iCWild [producers], 2018. Cape Town: DataFirst [distributor], 2018.10.25828/69td-0531
}
\keyword{adults}
\keyword{datasets}
\keyword{height}
\keyword{howell}
\keyword{kung}
\keyword{weight}
