% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_fertilisation-data.R
\docType{data}
\name{plant_fertilisation}
\alias{plant_fertilisation}
\title{Zea mays (maize) plant height in an experiment conducted by Darwin on
the effects of cross-fertilisation and self-fertilisation}
\format{A data frame with fifteen observations on two variables}
\source{
Darwin, C. R. 1878. The effects of cross and self fertilisation in the vegetable kingdom. 2d edition. London: John Murray, 16.
}
\usage{
plant_fertilisation
}
\description{
Measurements of the heights (in cm.) of different cross-fertilised and self-fertilised
 Zea mays (maize) plants in an experiment conducted by Darwin.
 Darwin details the experimental design on pages 11 to 15.
 The measurements appear orginally in inches (Darwin, 1878: 16)
 and are converted to centimetre and then rounded to one decimal place.
}
\details{
Zea mays (maize) plant height in an experiment conducted by Darwin on
the effects of cross-fertilisation and self-fertilisation.
}
\references{
Darwin, C. R. 1878. The effects of cross and self fertilisation in the vegetable kingdom. 2d edition. London: John Murray, 16.
Hand, D., Daly, F., McConway, K., Lunn, D., Ostrowski, E. (1993). A Handbook of Small Data Sets. New York: Chapman and Hall/CRC, 2.
}
\keyword{darwin}
\keyword{datasets}
