% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooiels-data.R
\docType{data}
\name{rooiels}
\alias{rooiels}
\title{Perceptions of Danger of Baboons in a small Cape Village of Rooiels}
\format{A data frame with cross-sectional and panel data.
Respondents are uniquely identified with the anonyous `respid` variable.
\describe{
  \item{respid}{Unique respondent identification number}
  \item{balanced_panel}{A variable which identifies if the same respondent was in both surveys}
  \item{year}{Year relating to the survey}
  \item{baboons_dangerous}{The result of a survey question: "Are the baboons in Rooeils dangerous at present"}
  }}
\source{
Centre for Social Science Research
}
\usage{
rooiels
}
\description{
This dataset contains a single question from a survey conducted in Rooiels a small coastal village
near Cape Town, South Africa.
The village lives in a sometimes difficult relationship with baboons.
Understanding how people percive baboons is important for thier management.
}
\references{
As source
}
\keyword{baboons}
\keyword{datasets}
\keyword{rooiels}
