% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons-data.R
\docType{data}
\name{seasons}
\alias{seasons}
\title{Traditional seasonal calander for Southern and Northern Hemisphere}
\format{tibble}
\source{
http://www.weathersa.co.za/learning/weather-questions/82-how-are-the-dates-of-the-four-seasons-worked-out
}
\usage{
seasons
}
\description{
On broad climatological and sociological grounds this data provides the months associated with the different seasons.
 An offical calender for the seasons does on astronomical grouds does not exist, so it should be enphasied that these divisions are somewhat crude.
 These dates are
}
\keyword{datasets}
\keyword{seasons}
