% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skulls-data.R
\docType{data}
\name{skulls}
\alias{skulls}
\title{Five time periods of measurements describing a skull in the Thebaid}
\format{A data frame with thirty skull realisaitons over five time periods for a total of 150 obsersavtions
\describe{
  \item{maximal_breadth}{Maximum length in millimeters}
  \item{basibregmatic_height}{Basibregmatic Height in millimeters}
  \item{basialveolar_length}{Basialveolar Length in millimeters}
  \item{nasal_height}{Nasal Height in millimeters}
  \item{time_period}{Time groupings}
  }}
\source{
Thomson, A. and Randall-Maciver, R. (1905) Ancient Races of the Thebaid, Oxford: Oxford University Press
The data is in Insert (which is at the end of the volume).
}
\usage{
skulls
}
\description{
to describe
}
\details{
Zea mays (maize) plant height in an experiment conducted by Darwin on
the effects of cross-fertilisation and self-fertilisation.
}
\references{
Hand, D.J., et al. (1994) A Handbook of Small Data Sets, New York: Chapman & Hall, pp. 299-301.
As cited in  Manly, B.F.J. (1986) Multivariate Statistical Methods, New York: Chapman & Hall.
}
\keyword{datasets}
\keyword{skulls}
