% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woodgate-display-data.R
\docType{data}
\name{woodgate_display}
\alias{woodgate_display}
\title{Woodgate's camera trap sighting for 25 sites in Karoo, South Africa}
\format{A tibble with 25 rows and 39 columns}
\source{
https://www.tandfonline.com/doi/suppl/10.2989/10220119.2018.1522513
}
\usage{
woodgate_display
}
\description{
Woodgate has 25 sites described in more detial in her paper.
The data is for a period of 2014 to 2016???
}
\details{
The data given in Table S2 of the supplementary resources.
}
\references{
Zoë Woodgate, Greg Distiller & Justin O’Riain (2018) Variation in mammal species richness and relative abundance in the Karoo, African Journal of Range & Forage Science,
35:3-4, 325-334, DOI: 10.2989/10220119.2018.1522513
}
\keyword{Karoo}
\keyword{cameratrap}
\keyword{datasets}
